+++ src/transformers/frameProcessors/index.ts
@@ -1,3 +1,3 @@
 export { processStreamsFrames } from './streamFrameProcessor';
 export { processMetricInstantFrames, processMetricRangeFrames } from './metricFrameProcessor';
-export { processHistogramFrames } from './histogramFrameProcessor';
+export { processHistogramFrames } from './histogramProcessors';