+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * BUGFIX: fix time range provided in `field_names` and `field_values` requests. Instead of being rounded to a 24-hour time range, the selected time range is now provided. See [pr #581](https://github.com/VictoriaMetrics/victorialogs-datasource/pull/581).
 * BUGFIX: fix template variables not being interpolated in the `Step` field of the query editor. See [pr #594](https://github.com/VictoriaMetrics/victorialogs-datasource/pull/594).
+* BUGFIX: fix the display of the histogram and bar chart for histogram stats functions. See [comment of issue #62](https://github.com/VictoriaMetrics/victorialogs-datasource/issues/62#issuecomment-3974623620).
 * BUGFIX: fix search of filter values in the query builder.
 
 ## v0.26.2